V3

+++ Ejercicio 1 .py
@@ -4,7 +4,7 @@ print('La edad de usted es',edad)
 print('Su peso es',peso,'Kg')"""
 
 #Ejercicio 1
-base=int(input("Ingrese la base del triagulo: "))
-altura=int(input('Ingrese la altura del triagulo: '))
+base=float(input("Ingrese la base del triagulo: "))
+altura=float(input('Ingrese la altura del triagulo: '))
 area=base*altura/2
 print("El area de su triagulo es",area)
